Add a Texture class with some coercion rules
If a parameter’s type is either a float or a
texture, we can decide which the user wants based
on the value type. Textures are always strings and
floats are always numbers.

Otherwise, it’s more difficult to decide and if we
can’t, then raise an error asking the user to
specify.